fixes type error

+++ WebApi/Controllers/AccountController.cs
@@ -55,7 +55,7 @@ namespace WebApi.Controllers
 
         // GET: api/Account/5
         [HttpGet("{id}", Name = "GetAccount")]
-        public async Task<ActionResult<UserDto>> Get(int id)
+        public async Task<ActionResult<UserDto>> Get(long id)
         {
             var user = await _context.User.FindAsync(id);
             if (user == null)
